chore:17.0 change type

+++ 0x17-doubly_linked_lists/lists.h
@@ -10,7 +10,7 @@
  */
 typedef struct dlistint
 {
-	size_t n;
+	int n;
 	struct dlistint *prev;
 	struct dlistint *next;
 } dlistint_t;
